Implemented spawn pools that can spawn items or monsters

+++ core/src/com/mygdx/game/Inventory/ItemTypes/HeadItem.java
@@ -37,5 +37,5 @@ public abstract class HeadItem implements InventoryItem {
         return itemContextActions;
     }
 
-    public abstract int getDefenceRating();
+    public abstract float getDefenceRating();
 }
